菜单前端验证 后盾接口开发

+++ share-basic/share-dao/src/main/java/org/andot/share/basic/entity/Organ.java
@@ -6,7 +6,7 @@ import lombok.EqualsAndHashCode;
 @Data
 @EqualsAndHashCode(callSuper = false)
 public class Organ extends BaseEntity {
-    private Integer organId;
+    private Long organId;
     private String organName;
     private Byte organType;
     private Byte orderCode;
